Don't abort early when we get a timeout for *previously* solved programs
If we get a timeout at the first request, it means that the program was
completed earlier (successfully or not), so we can continue to the next program.

If we get a timeout at a later request, we know it's our fault and that we run
into a limitation, so we better stop, unless we're rushing - add a flag for the
"rushing" mode.

Note: The structure is messy, to refactor soon.